requirements.txt
discord.py
emoji
colorama
motor